moved around files to fix client vs app settings
