+++ app/configuration.py
@@ -10,17 +10,16 @@ class Config(object):
 	"""
 	DEBUG = False
 	TESTING = False
-	DATABASE_URI = 'sqlite:///application.db'
 	BOOTSTRAP_FONTAWESOME = True
-	SECRET_KEY = "MINHACHAVESECRETA"
+	SECRET_KEY = "MINHACHAVESECRasdETA"
 	CSRF_ENABLED = True
 
 	#Get your reCaptche key on: https://www.google.com/recaptcha/admin/create
 	#RECAPTCHA_PUBLIC_KEY = "6LffFNwSAAAAAFcWVy__EnOCsNZcG2fVHFjTBvRP"
 	#RECAPTCHA_PRIVATE_KEY = "6LffFNwSAAAAAO7UURCGI7qQ811SOSZlgU69rvv7"
 
-class ProductionConfig(Config):
-	DATABASE_URI = 'mysql://user@localhost/foo'
+# class ProductionConfig(Config):
+# 	DATABASE_URI = 'mysql://user@localhost/foo'
 
 class DevelopmentConfig(Config):
 	DEBUG = True
