+++ client.py
@@ -13,5 +13,4 @@ from client import client
 #----------------------------------------
 
 if __name__ == "__main__":
-	port = int(os.environ.get("PORT", 3000))
-	client.run(host='0.0.0.0', port=port)
+	client.run(host='0.0.0.0', port=3000)
