+++ client/templates/base.html
@@ -1,0 +1,48 @@
+{% extends "bootstrap_base.html" %}
+
+{%block title%}Base{%endblock%}
+
+{%block body_content%}
+
+<div class="container">
+	<div class="row">
+		<div class="span12">
+
+			<div class="navbar">
+			  <div class="navbar-inner">
+			    <a class="brand" href="#">MySiteName</a>
+			    <ul class="nav">
+			      <li class="active"><a href="/">Home</a></li>
+			      <li><a href="/list">List</a></li>
+			      <li><a href="/new">New</a></li>
+			    </ul>
+			  </div>
+			</div>
+
+			<form class="form form-inline pull-right">
+			  <input type="text" class="span2">
+			  <button type="submit" class="btn">Submit</button>
+			</form>
+
+		</div>
+	</div>
+
+    {% with messages = get_flashed_messages() %}
+	    {% if messages %}
+	    <div class="row">
+	    	<div class="span8 offset4">
+			    <ul>
+			    {% for message in messages %}
+			        <li>{{ message }} </li>
+			    {% endfor %}
+			    </ul>
+			</div>
+		</div>
+	    {% endif %}
+    {% endwith %}
+
+	{%block main_content%}{%endblock%}
+
+</div>
+
+{%endblock%}
