+++ client/templates/index.html
@@ -1,0 +1,31 @@
+{% extends "base.html" %}
+
+{%block title%}Index!{%endblock%}
+
+{%block main_content%}
+<script src = "https://cdnjs.cloudflare.com/ajax/libs/axios/0.16.2/axios.js"> </script>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="static/main.js"></script>
+
+	<h1>Index</h1>
+	<div class="row">
+		<div class="span4">
+			<img src="http://placehold.it/360x200" class="img-polaroid">
+		</div>
+		<div class="span4">
+			<img src="http://placehold.it/360x200" class="img-polaroid">
+		</div>
+		<div class="span4">
+			<img src="http://placehold.it/360x200" class="img-polaroid">
+		</div>
+	</div>
+
+	<input id="input">
+
+	<input type="submit" id="submit">
+
+	<p id="test"></p>
+
+
+
+{%endblock%}
