+++ client/templates/list.html
@@ -1,0 +1,61 @@
+{% extends "base.html" %}
+{% import "bootstrap_wtf.html" as wtf %}
+
+{%block title%}Lista!{%endblock%}
+
+{%block main_content%}
+
+	<h1>Listagem</h1>
+	
+	<div class="row">
+		<div class="span12">
+
+			Listagem...
+
+			<table class="table">
+				<tr>
+					<th>Coluna 1</th>
+					<th>Coluna 2</th>
+					<th>Coluna 3</th>
+					<th>Coluna 4</th>
+					<th>Coluna 5</th>
+					<th>Coluna 6</th>
+				</tr>
+				<tr>
+					<td>valor 1</td>
+					<td>valor 2</td>
+					<td>valor 3</td>
+					<td>valor 4</td>
+					<td>valor 5</td>
+					<td>valor 6</td>
+				</tr>
+				<tr>
+					<td>valor 1</td>
+					<td>valor 2</td>
+					<td>valor 3</td>
+					<td>valor 4</td>
+					<td>valor 5</td>
+					<td>valor 6</td>
+				</tr>
+				<tr>
+					<td>valor 1</td>
+					<td>valor 2</td>
+					<td>valor 3</td>
+					<td>valor 4</td>
+					<td>valor 5</td>
+					<td>valor 6</td>
+				</tr>
+				<tr>
+					<td>valor 1</td>
+					<td>valor 2</td>
+					<td>valor 3</td>
+					<td>valor 4</td>
+					<td>valor 5</td>
+					<td>valor 6</td>
+				</tr>
+			</table>
+
+		</div>
+	</div>
+
+{%endblock%}
