+++ client/templates/login.html
@@ -1,0 +1,27 @@
+{% extends "base.html" %}
+{% import "bootstrap_wtf.html" as wtf %}
+
+{%block title%}Login!{%endblock%}
+
+{%block main_content%}
+
+	<h1>Login</h1>
+	<div class="row">
+		<div class="span12">
+
+			<form action="." method="post" name="loginform">
+				{{ wtf.form_errors(form,"only") }}
+
+				{{ form.hidden_tag() }}
+				{{ wtf.horizontal_field(form.user) }}<br>
+				{{ wtf.horizontal_field(form.password) }}
+
+				<div class="form-actions">
+					<button name-"action_save" type="submit" class="btn btn-primary">Salvar</button>
+				</div>
+			</form>
+
+		</div>
+	</div>
+
+{%endblock%}
