+++ client/templates/new.html
@@ -1,0 +1,29 @@
+{% extends "base.html" %}
+{% import "bootstrap_wtf.html" as wtf %}
+
+{%block title%}Novo!{%endblock%}
+
+{%block main_content%}
+
+	<h1>Novo</h1>
+	<div class="row">
+		<div class="span12">
+
+			<form action="/save/" method="post" name="exampleform">
+				{{ wtf.form_errors(form,"only") }}
+
+				{{ form.hidden_tag() }}
+				{{ wtf.horizontal_field(form.title) }}<br>
+				{{ wtf.horizontal_field(form.content) }}
+				{{ wtf.horizontal_field(form.date) }}
+				{# wtf.horizontal_field(form.recaptcha) #}
+
+				<div class="form-actions">
+					<button name-"action_save" type="submit" class="btn btn-primary">Salvar</button>
+				</div>
+			</form>
+
+		</div>
+	</div>
+
+{%endblock%}
